\name{DSBayes-package}
\alias{DSBayes-package}
\alias{DSBayes.Intro}
\docType{package}

\title{Bayesian subgroup analysis in clinical trials}

\description{Calculate posterior modes and credible intervals of parameters of the Dixon-Simon model for subgroup analysis (with binary covariates) in clinical trials.}

\details{
\tabular{ll}{
Package: \tab DSBayes\cr
Version: \tab 1.1\cr
Date: \tab Dec 27, 2013\cr
Depends: \tab R (>= 2.15.1)\cr
Imports: \tab BB \cr
License: \tab GPL Version 2 \cr
}


The main functions in this package are:
\preformatted{

DSBayes:  A function to calculate the posterior mode and credible interval of the parameters in the Dixon-Simon model, as well as the MLE regression coefficients. 
}
}

\author{Ravi Varadhan <rvaradhan@jhmi.edu> and Wenliang Yao <yaow080@gmail.com>}


\references{
   Dixon D. and Simon R. (1991).  Bayesian Subset Analysis. \emph{Biometrics}, 47, 871-881
   }
   
\keyword{ package }
 